+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/ImageNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class ImageNotExistException extends NotExistException {
-    public ImageNotExistException(Long imageId) {
+    public ImageNotExistException(Integer imageId) {
         super(String.format("Image %d not exist", imageId));
     }
 
-    public ImageNotExistException(Long imageId, Throwable cause) {
+    public ImageNotExistException(Integer imageId, Throwable cause) {
         super(String.format("Image %d not exist", imageId), cause);
     }
 }
