+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/ImageToProductNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class ImageToProductNotExistException extends NotExistException {
-    public ImageToProductNotExistException(Long imageToProductId) {
+    public ImageToProductNotExistException(Integer imageToProductId) {
         super(String.format("Image to product %d not exist", imageToProductId));
     }
 
-    public ImageToProductNotExistException(Long imageToProductId, Throwable cause) {
+    public ImageToProductNotExistException(Integer imageToProductId, Throwable cause) {
         super(String.format("Image to product %d not exist", imageToProductId), cause);
     }
 }
