+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/OrderNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class OrderNotExistException extends NotExistException {
-    public OrderNotExistException(Long order) {
+    public OrderNotExistException(Integer order) {
         super(String.format("Order %d not exist", order));
     }
 
-    public OrderNotExistException(Long order, Throwable cause) {
+    public OrderNotExistException(Integer order, Throwable cause) {
         super(String.format("Order %d not exist", order), cause);
     }
 }
