+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/ProductNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class ProductNotExistException extends NotExistException {
-    public ProductNotExistException(Long productId) {
+    public ProductNotExistException(Integer productId) {
         super(String.format("Product %d not exist", productId));
     }
 
-    public ProductNotExistException(Long productId, Throwable cause) {
+    public ProductNotExistException(Integer productId, Throwable cause) {
         super(String.format("Product %d not exist", productId), cause);
     }
 }
