+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/ShipperNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class ShipperNotExistException extends NotExistException {
-    public ShipperNotExistException(Long shipperId) {
+    public ShipperNotExistException(Integer shipperId) {
         super(String.format("Product %d not exist", shipperId));
     }
 
-    public ShipperNotExistException(Long shipperId, Throwable cause) {
+    public ShipperNotExistException(Integer shipperId, Throwable cause) {
         super(String.format("Product %d not exist", shipperId), cause);
     }
 }
