+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/UserNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class UserNotExistException extends NotExistException {
-    public UserNotExistException(Long userId) {
+    public UserNotExistException(Integer userId) {
         super(String.format("User %d not exist", userId));
     }
 
-    public UserNotExistException(Long userId, Throwable cause) {
+    public UserNotExistException(Integer userId, Throwable cause) {
         super(String.format("User %d not exist", userId), cause);
     }
 }
